Sync with libnetconf2-v4.1, update to the latest releases
The YANG models have changed, so this requires a data migration:

- the UNIX socket path is now hidden (and upstream has added more
  config options that we want to disable for security reasons),

- the IP address that we're binding to has moved to a different schema
  path.

This changes some data which were previously done in an earlier
migration step, which means that we can drop some of these changes and
delegate them until the very last step (for v13).

Our Vuildroot changes are available in the `cesnet/2025-12-04` branch.

Change-Id: I2bcda81e01e1a6537fce9c0d94e96b1483220378
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/9026
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/9032
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/netconf-cli/+/9027