tools: add simple tests to eeprom content generator
Unfortunately, python's doctest test specification is a little
cumbersome for byte strings. If the byte is a printable char, it expects
the character. If it is not a printable char, it expects the hex-escaped
version like \\xab.
I did not like mixing both styles in the same test as it looked weird to
have strings like "\\xab(". Let's rather use the output of hex() method.

Change-Id: I7f353357465eb9c809d575deefa7c41ae974b96d